+++ app/src/main/java/ru/pg13lac/nbanews/presentation/ui/more/MoreFragment.kt
@@ -1,8 +1,8 @@
-package ru.pg13lac.nbanews.presentation.ui.settings
+package ru.pg13lac.nbanews.presentation.ui.more
 
 import ru.pg13lac.nbanews.R
 import ru.pg13lac.nbanews.presentation.ui.base.BaseFragment
 
-class SettingsFragment : BaseFragment() {
+class MoreFragment : BaseFragment() {
     override val layoutRes = R.layout.fragment_settings
 }
