+++ app/src/main/res/layout/fragment_settings.xml
@@ -5,7 +5,7 @@
     android:id="@+id/mainConstraint"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.settings.SettingsFragment">
+    tools:context=".presentation.ui.more.MoreFragment">
 
     <TextView
         android:id="@+id/tvSettings"
