+++ app/src/main/res/navigation/settings.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/settingsFragment"
-        android:name="ru.pg13lac.nbanews.presentation.ui.settings.SettingsFragment"
+        android:name="ru.pg13lac.nbanews.presentation.ui.more.MoreFragment"
         android:label="fragment_settings"
         tools:layout="@layout/fragment_settings" />
 </navigation>
